+++ modules/properties/ui/src/sap/commerce/toolset/properties/ui/CxPropertyMouseHandler.kt
@@ -92,6 +92,7 @@ internal class CxPropertyMouseHandler(
 
         return CxPropertyRowAction.entries
             .find { fromRightEdge <= JBUI.scale(CxPropertyRowAction.offsetOf(it)) }
+            ?.takeIf { list.editable || !it.remote }
     }
 
     companion object {